+++ example/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@emotion/react": "^11.4.0",
     "@emotion/styled": "^11.3.0",
+    "@hookform/resolvers": "^2.8.1",
     "@kyuzan/mint-sdk-js": "latest",
     "@reduxjs/toolkit": "^1.5.1",
     "@tippyjs/react": "^4.2.5",
@@ -40,7 +41,8 @@
     "redux": "^4.1.0",
     "redux-logger": "^3.0.6",
     "slick-carousel": "^1.8.1",
-    "vercel": "^22.0.1"
+    "vercel": "^22.0.1",
+    "yup": "^0.32.9"
   },
   "devDependencies": {
     "@babel/core": "^7.13.14",
