+++ example/src/components/organisms/ShippingInfo/index.tsx
@@ -1,0 +1,1 @@
+export { Presentation as ShoppingInfo } from './presentation'
